Added a utility to load a Scriban template from an embedded resource
